Comment out API stuff; add vertical scrolling

+++ base/plugins/scenario/source/Document/Constraint/ViewModels/AbstractConstraintView.cpp
@@ -14,13 +14,13 @@ void AbstractConstraintView::setInfinite(bool infinite)
     update();
 }
 
-void AbstractConstraintView::setDefaultWidth(int width)
+void AbstractConstraintView::setDefaultWidth(double width)
 {
     prepareGeometryChange();
     m_defaultWidth = width;
 }
 
-void AbstractConstraintView::setMaxWidth(bool infinite, int max)
+void AbstractConstraintView::setMaxWidth(bool infinite, double max)
 {
     prepareGeometryChange();
 
@@ -33,13 +33,13 @@ void AbstractConstraintView::setMaxWidth(bool infinite, int max)
 
 }
 
-void AbstractConstraintView::setMinWidth(int min)
+void AbstractConstraintView::setMinWidth(double min)
 {
     prepareGeometryChange();
     m_minWidth = min;
 }
 
-void AbstractConstraintView::setHeight(int height)
+void AbstractConstraintView::setHeight(double height)
 {
     prepareGeometryChange();
     m_height = height;
